Inventory and Shop
Added Gather Button functionality

+++ Assets/Project/Scripts/Inventory/InventorySO/InventorySO.cs
@@ -3,6 +3,6 @@
 [CreateAssetMenu (fileName = "Inventory", menuName = "ScriptableObjects/Inventory/Inventory")]
 public class InventorySO : ScriptableObject
 {
-    public int inventoryWeight;
+    public float inventoryWeight;
     public InventoryItemView inventoryItemView;
 }
